requirements.txt
Flask>=2.2
Flask-SQLAlchemy
Flask-Migrate
python-dotenv